+++ linux/sys/cpuvar.h
@@ -116,7 +116,7 @@ typedef struct cpu_trap_t {
 
 typedef struct cpu_core {
         uint16_t        cpuc_dtrace_flags;      /* DTrace flags */
-	uint8_t         cpuc_dcpc_intr_state;   /* DCPC provider intr state */
+	uint32_t        cpuc_dcpc_intr_state;   /* DCPC provider intr state */
 	uint8_t		cpuc_probe_level;	/* Avoid reentrancy issues in dtrace_probe */
 	uint32_t	cpuc_this_probe;	/* Current probe.	*/
 //	spinlock_t	cpuc_spinlock;
